Backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
bcrypt
requests